requirements.txt
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
colorama==0.4.6
et_xmlfile==2.0.0
execnet==2.1.2
h11==0.16.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
pycparser==2.23
Pygments==2.19.2
PySocks==1.7.1
pytest==9.0.1
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-xdist==3.8.0
PyYAML==6.0.3
selenium==4.38.0
setuptools==80.9.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
wheel==0.45.1
wsproto==1.3.1
